CORE-1734 Add Terminate action to Analyses Listing rows
