+++ src/components/analyses/listing/Actions.js
@@ -13,18 +13,20 @@ import ids from "../ids";
 import buildID from "components/utils/DebugIDUtil";
 
 import {
-    isInteractive,
+    allowAnalysesCancel,
     allowAnalysisTimeExtn,
     isBatchAnalysis,
-    useRelaunchLink,
-    useGotoOutputFolderLink,
+    isInteractive,
     isTerminated,
+    useGotoOutputFolderLink,
+    useRelaunchLink,
 } from "../utils";
 
 import { useConfig } from "contexts/config";
 
 import { Grid, IconButton } from "@material-ui/core";
 import {
+    Cancel as CancelIcon,
     HourglassEmptyRounded as HourGlass,
     Launch as LaunchIcon,
     PermMedia as OutputFolderIcon,
@@ -92,6 +94,7 @@ export default function Actions(props) {
         allowBatchDrillDown = true,
         handleDetailsClick,
         handleInteractiveUrlClick,
+        handleTerminateSelected,
         handleBatchIconClick,
         setPendingTerminationDlgOpen,
         baseId,
@@ -107,6 +110,7 @@ export default function Actions(props) {
 
     const isBatch = isBatchAnalysis(analysis);
     const isVICE = isInteractive(analysis);
+    const allowCancel = allowAnalysesCancel([analysis], username, config);
     const allowTimeExtn = allowAnalysisTimeExtn(analysis, username, config);
     const [relaunchHref, relaunchAs] = useRelaunchLink(analysis);
     const [outputFolderHref, outputFolderAs] = useGotoOutputFolderLink(
@@ -116,6 +120,16 @@ export default function Actions(props) {
 
     return (
         <Grid container direction="row" wrap="nowrap">
+            {allowCancel && (
+                <IconButton
+                    size="small"
+                    onClick={handleTerminateSelected}
+                    id={buildID(baseId, ids.TERMINATE_BTN, ids.BUTTON)}
+                    title={t("terminate")}
+                >
+                    <CancelIcon color="error" fontSize="small" />
+                </IconButton>
+            )}
             {isTerminatedAnalysis && (
                 <Link href={outputFolderHref} as={outputFolderAs} passHref>
                     <GotoOutputFolderButton
